+++ pybamm/expression_tree/independent_variable.py
@@ -40,9 +40,9 @@ class IndependentVariable(pybamm.Symbol):
     def to_equation(self):
         """Convert the node and its subtree into a SymPy equation."""
         if self.print_name is not None:
-            return sympy.symbols(self.print_name)
+            return sympy.Symbol(self.print_name)
         else:
-            return sympy.symbols(self.name)
+            return sympy.Symbol(self.name)
 
 
 class Time(IndependentVariable):
@@ -71,6 +71,10 @@ class Time(IndependentVariable):
         See :meth:`pybamm.Symbol.evaluate_for_shape()`
         """
         return 0
+
+    def to_equation(self):
+        """Convert the node and its subtree into a SymPy equation."""
+        return sympy.Symbol("t")
 
 
 class SpatialVariable(IndependentVariable):